i18n support requires separate toolNames to get unique displayNames
Dataverse will look for local-specific displayNames in 'bundle files' -
the lookup is by toolName so only one label can be defined per toolName.

AFAIK this is the first functionality where having toolname and
displayname not be 1-to-1 is a problem.

Note - the same toolname can still be used for multiple mimetypes (as
long as they share a common displayName like 'Play Audio')